Version 3.2.0 of WoW Raid Manager
Name change from phpRaid to WoW Raid Manager (WRM for short).  This
was the security update version as well as implimentation of the full
GPL 3.0 License (GPL License was updated from GPL2 to GPL3).  By and
large most of the code changes in this release were to scrub the input
variables from the users and to properly ensure that all database queries
had scrubbed and sanitized data going to them.