General fix to E107/WRM Auth, bug fixing of a few bugs.
This is a general update to the e107 authentication / installation files.
This update fixes several issues coresponding to e107 and WRM linkage.

The two major issues fixed are:

* WRM Admins who install WRM with the e107 authentication option who's
E107 user id is NOT 1 (i.e. The admin account selected in WRM is not
User ID 1) cannot get admin permissions when logging in after
installation.

* Profiles created after installation fo 3.2.0 are created blank.  Users
with these profiles have problems creating characters, signing up for
raids, and other basic usage of WRM.

Several bug fixes in the code.  In the installation file the user id of
the admin user created was always assumed to be 1.  The insert statement
placed a "0" into the profile ID field relying upon the "AutoNumber" to
properly fill in the profile ID, causing the Admin ID to always be 1 in
the WRM table, which did not match the e107 ID under many conditions.

In the authentication file the code to create a new user if properly
authenticated to e107 but without a profile in the WRM tables was
incorrectly using the wrong string format for many of the fields causing
a blank profile to be created.