Patch 7: Visual Navigation to Left Side Menu, fix to RSS.php
This patch adds new functionality to WRM and fixes a minor bug with the
RSS feed.

Users using the RSS feed to publish their raids to websites and web pages
notice that raids marked "old" in the system by raid leaders and officers
still show up in the RSS feed.

This patch also adds visual navigation to the left side menu of WRM
(Thanks to LuL0r for the code and CSS information for this). When users
click a left side menu "button" the button now appears differently to
indicate which screen the users are on.