Unable to Display error for Username or email not matching.
Patch contributed by Diska of wowraidmanager.net.  The current
e107 installation contains code to check the username/email
address against that of the e107 database before installing
the admin user.  If the username/email check fails there is
supposed to be an "or die" run by the code.  However due to
the way php works this "or die" would only check for a syntax
error in the SQL code, NOT for the condition where the user
and email don't match.  Patch fixes this so that the error for
user/email doesn't match gets displayed correctly.