Gender Support from patch 2 still didn't fix gender problem.
Congtributed by LuL0r from wowraidmanager.net.  The original
code from LuL0r had an issue in it.  When the user went back
and switched languages from english to german or visa versa
more problems with the gender came up.  Patch 2 is a valid fix
till a user goes back to make changes to his or her language.

This patch fixes this.  Instead of trying to match the language
files for gender name, it simply forces gender internally to
"male" or "female" and only uses the language files for
display.