Added a utility to package up a release and manage permissions.
Added a new utility that goes through the directory structure
changes all of the permissions to appropriate unix permissions
(i.e. no execute on text files and such) and then goes through
the process of creating a release in the "package" directory.

At the end of this process two files are created, one .tar.gz
the other a .zip file with the name of the release.

This takes 1 parameter: version number specified in dotted
format, number only (3.2.1, not v3.2.1).