Implimentation of WoW Armory Character Popups
Implimented the WoW Armory Character Popups for all areas of the WRM code.
Anyplace a character name used to appear it should now appear as a link
to the WOW Armory and when moused over, should contain a popup box that
contains relivant information about the character as he or she exists on
the armory at the moment.

All configure options are set in the configuration section of WRM including
server name, wow armory link, armory language code, etc.