JavaScript Date Creation: Issue with date sort fix.
The following code is temporary.  This code causes the report generator
to call a function to generate a date based upon a javascript call, instead
of the current functionalty that generates a date.  The javascript has been
changed to support German and US dates in a specific format but this needs
to be made more generic.