Updates to Multiple Signup Code, Small Language Changes
This generic update handles two changes: first it enables the display of
multiple character signup on one account, i.e. if one person signs up for
a raid with two characters (assuming the options to do so are checked)
there will be an identification of such.

Secondly there were a couple small changes to the german and english
language files.  Contributed by Nalumis.