Partial Implementation of Screen Width
Removed the 875 px hard coded width of the center colum and allowing it to
float now based upon the size of the screen and the size specified in the
css file.  This is half a fix contributed by Nalumis that is intended to
lengthen the raid screen to full screen size.  The other half of the fix
is in templates/SpiffyJr/style/stylesheet.css line 221.  The 1000px width
needs to be removed and a percentage added (95% for full screen).