WRM 3.2.1 Patch 1
Fixes an issue in the install.sql file where the armory configuration items
were not placed into the config table upon installation, causing armory
lookup not to work for many new installations.

Fixes an issue in the 3.2.1.sql file that had my table prefix (_TC_) in
the sql causing most people to fail the application of this patch.