BugFix: New Raid button doesn't show up if no raids in database.
V3.5 Introduced an error where the "new raid" button does not show up if
there are no current raids in the database.  This was due to the placement
of the new code in 3.5 to generate the new raid signup button, in previous
versions the button was in the HTML file, not generated through code.