Initial Push of EqDKP Integration from Neme
Changes the base code to support integration with an EqDKP installation.
The changes here create a new side menu item entitled "DKP" that pulls up
information from the DKP system and displays it in WRM.  Users are linked
back into the DKP system to provide direct access to the DKP information.

Several changes were made from the base release of this on the suggestions
form including configuration items to turn this on and off as well as info
for how to access the base DKP Install.  More changes to this comming, this
is NOT a final release for the DKP integration.