Added Configuration Items in the Config Area.
Three new config items have been added to the configuration area and a new
configuration section "Integration with External Applications".  The new
section is designed to hold config items appropriate to linking WRM with
other non-wrm applications.

The config items added were: "Enable Armory Lookups" to turn on or off the
3.2.1 armory lookup.  "Enable EqDKP Integration" to enable integration with
the EqDKP DKP system.  And finally, "EqDKP URL" to tell the system where
to point to to link into the EqDKP installation for things like user
pages and base links.