Removal of config.php from distribution.
The config.php file will no longer be distributed with the system.  This
file is autoatically created at installation time or is now (new) read in
at upgrade time and re-produced as is.  This should illeviate upgraders
from accidently overwriting the old config file as well as preserve database
information for "other" databases, such as EqDKP.