Date Fix: International Dates (re-write)
This is a total re-write of the international date sort fix and handling
all the rest of the date issues in the code.  This should be a final
implementation of date code.

The current date / time configuration in the config area now only changes
the DISPLAY of the date/time.  All date/times are passed into the report
function as the ISO YYYY/MM/DD HH:mm:ss format and thus is sortable.  The
report class has also been changed to support three new date formats:

wrmdate - Format the date output according to the date format from config.
wrmtime - Format the date output according to the time format from config.
wrmdatetime - Format the date output according to the date and time format.

All Date Sorting and Display issues should be fixed as of this version.