Schedule Raids in 12/24 Hour Format.
Allow scheduling of raids in 12/24 hour format.  A switch was put into the
configuration section that allows raids to be scheduled in 12 or 24 hour
format.  This option is standalone from the DISPLAY of the date data.  Thus
raids could be scheduled in 12 hour (am/pm) format but DISPLAYED in 24 hour
format, or visa versa.