Bug 0000031: Text for Title and Message headers do not show up.
There was an issue with the langauge files for all languages that had the
wrong keys for announcements_title_text and announcements_messages_text.

The keys were announcements_title_txt and announcements_msg_txt in the
language file and as above in the code.

Note: Italian is not included in this patch due to there being other
changes for the italian language files.