Calendar Re-Write!
This is the re-write of the WRM calendar from the ground up.  The calendar
now displays as a calendar instead of a running list of days/weeks.  There
are searchable functions in teh calendar to pull up a month/year of choice
and a next/previous button to scroll between months.  The calendar now uses
a .htm file for it's formatting and the standard WRM template classes.

The language updates are all calendar related.