New Feature: Password Change Ability
Added a new feature (originally submitted by Neme) allowing for password
changes.  There are two password change abilties.  The first is an "I
forgot my password" change ability and the second is a "Change password"
ability.  Users now have the ability to change a forgotten password
(a random password is emailed to them) and to change a password to a new
password.