Bug: 0000056 - Users.php problems
Fixed a couple bugs related to deleting users.

The first is that the username was never set properly ($data['name']
was used instead of $data['username']) so the delete messages passed to
the log all did not contain the deleted username.

Secondly users deleted with active signups are deleted leaving the
signups sitting around. This causes problems with outputting raids to
LUA and other issues.