Smarty Update to Raids.php
This is an "in progress" update to the raids display part of WRM.  At the
moment smarty is implemented for the initial raids display, columns are
pulled from a new table and the display is looking very close to the old
functionality.  I still have data sorting and record pagination left.