Bugfix for Hack attempt on page 2+ of missing signups view.
This is a fix for a hack attempt being reported when the user clicks page
2 or above of the "Missing Signups" report accessed from the raid view
page.  The identified problem was that the link was not passing along the
"raid_id" and instead passed along "mode=view" which has no use on the
missing signups table.  Because raid_id was not passed the code determined
this as a hacking attempt.