Added support for users in alterante groups being allowed to use WRM.
When looking at group access restrictions (set by the installation routine)
the SMF and SMF2 bridges were ONLY looking at the primary group (that
identified by the ID_GROUP field) and not the secondary/additional groups
identified by the additionalGroups field.  When a user is in a primary
group that is not given access to WRM but a secondary group that is, they
are still denied access to WRM.  This has been fixed by this patch.