Mysql Connection Bugfix
Upon creation of a connection to a server a database is selected.  When
opening another connection to a different database the application
re-selects the database but does not open a new connection (server, user,
and password are the same so it re-uses the old connection), thus the
database is now changed and other queries are now using the wrong database.

A new parameter was added to the sql_db code to allow for a new connection.
Passing "TRUE" after the database name in the sql_db (before the
persistency parameter) allows for a new connection to a different database
to be made.