Update for Location selecting Events.
This should be a final implementation of a major update to Locations.  The
location page is now able to schedule WoW Major Events.  An events table
has been created holding all of WoW's official zones as well as the PvP,
Meeting, and "Other" events.  Locations now links with this via pulldown
menus to populate the available WoW Events.