Fix for Double Quote (") in announcements.
Putting a " in an announcement caused the announcement not to be allowed
to delete.  The fix was simply to remove the title pass in to the delete
functionality, delete only needs the announcement ID, it doesn't need the
title as well.  This also fixes a security bug where a user could put a "
into an announcement and follow it up with a set for something like
phpraid_dir to direct users off site.