Various fixes for Internet Explorer:
* Icon for unselected raids turns up as a red X (image that is blank shows
as an invalid image in IE).  This gets fixed if you select an image but
the default image when the screen loads is the red X missing image.

*When selecting the blank from event name dropdown menu (i.e. removal of
prior selection), IE loads the index.php screen instead of reloading the
current screen.  The fix was to add location.php URL to the "blank"
dropdown's value to tell it to reload the current screen instead of the
index screen.