Fixed a bug with WRM Lua Output where foreign languages would not properly output.
The original lua output was expecting to match class names against the
foreign language equivalent of a class name.  In 4.0 all classes internally
are US names for classes and are only output as foreign class names on
the DISPLAY.  Thus players using foreign languages where the class names
were different than the english names would cause drafted characters not to
show up on the lua output.

Note: There my still be a bug with output of LUA with the class names
always showing in english instead of the output DATA being in the foreign
language.