Fixes a bug where the WRM header was ALWAYS being displayed, even in cases where the header was already displayed.
So for instance when an error message was occuring, the page would load
the header would print, the error message would print and then the page
would be shown AGAIN due to the header being displayed.

This change forces the header ONLY to get displayed when there are no
errors.

This change may be rolled back in a later release if there's a better way
to do this.