Update to add part of the Permissions section to the Admin Area.
This part of the code adds the front page of permissions (the ability to
edit and modify permission groups and the permissions those groups contain)
and to create new permission groups.

The functionality to modify the details of the permission group and who
is in it has been removed and will be re-written into the user management
area of the Admin Interface.