Adding Raid Force information to calendar, index, locations, raids and view to allow for viewing and updating of Raid Force.
This should be the final code for raid force other than the fact that raid
force is not automatically set from the location dropdown in the raid
create area.  This is a bug that needs fixed.

Lastly, LUA Output needs to be modified to support pulling only the
information from a particular raid force.