Changes to Step0 and Step1.  Step0 changes include changing the common file from common.php to install_common.php, changing the relative pathing to the wrm Config File to an explicit calculated path, and adding the new World_of_Warcraft.sql file to the installer's knowledge.
Step1 changes include checking more directories in the directory checking
code (instead of just checking wrm/cache we're now checking most of the
directories under the cache directory), moving the PHP Version Check down
below the file permissions check (so that the code for the two isn't mixed
together), and fixing a bug with the config file check.

In the config file check area, it was assumed (with the old code) that if
the config file didn't already exist (is_file returned False) that the file
could properly be written to.  However the base WRM directory was never
checked for write ability in the dir check section, thus causing an issue
where if the WRM directory wasn't owned by the person running the web
server the installation would fail somewhere around step4, even when
reporting no problems in step 1.

Finally in step 7 we've added code to add the new game file into the
database.

Signed-off-by: Illydth <douglasw@wagnerweb.org>