Implementation of a new feature:  Limitation of the number of "old" raids listed on the index page.
There is a new option in the General Configuration for WRM that allows a
user to identify the number of "old" raids that the index page will show.

That limit is then applied in a query that pulls all "current" and only
that number of "old" records from the database when the index page is
rendered.

This allows tons of old records to be kept around but not affect page load
times.

Signed-off-by: Illydth <douglasw@wagnerweb.org>