Added code to create a return link to the view.php page the user came from.
Also added code to validate users characters.  Instead of just checking
if the profile doesn't exist in signups, the missing_signups.php now checks
to see if those profiles have valid characters that COULD be signed up to
the raid before flagging them as "missing."

In other words, if I have a profile and character but he's level 50, my
profile won't show up in the list of missing profiles for a level 60 raid
but will show up as missing for signups to a level 45 - 55 instance run.

Signed-off-by: Illydth <douglasw@wagnerweb.org>