Initial try at automatic raid marking as old from code sent to me by Istari.
This code sets up the Scheduler/Automatic Executor to pull a list of all
open raids (non marked recurring and not marked old) in the system and
check the current date/time against the start date of the raid (plus a fudge
factor determined by the WRM admin configuration).  If this time is passed,
the system will update the raid "old" mark to 1, thereby automatically
making it an "old" raid.

At this point all is complete from the mark side, I need to add the config
items for turning the feature on/off and for setting the "fudge factor"
before the raid gets marked.

Signed-off-by: Douglas Wagner <douglasw@pixie.wagnerwebhome.org>