Standardized Page URL code. Changed "Tag" to a more appropriate name (Event Type). Fixed broken code related to Editing a Raid.  Currently when a raid is edited, the EventID is dropped, causing the raid to become unlinked from the location and causing the raid icon (among other things) not to show properly on the calendar.
MUCH was missing/broken with the code.  No name was being assigned to the
array variable, $raid_location (what is being compared against to determine
what is "selected" doesn't even exist in this section of code, and the
entire options array was missing from being passed into the template.

All of these are now fixed.

Signed-off-by: Douglas Wagner <douglasw@pixie.wagnerwebhome.org>