Cookie/Login fix:
Problem identified with cookies and being logged out of the software
tracked back to an issue with capitalization.  The incoming cookie username
was NOT being passed through strtolower, and thus did not match the
username being passed in or compared against (which WAS being strtolowered).

Thanks go dn on the WRM boards for this.

Signed-off-by: Douglas Wagner <douglasw@pixie.wagnerwebhome.org>