Identified by Mada: Issue with functions_users.php showing "signups" for main page.
Because we're specfiying "queue = 0" explicity, only those users in
"drfated" status are being picked up as "signed up to the raid" on the
front page.  Instead if we removed the "queue" check altogether (queue=1
means the user is Queued, queue=0 means the user is not queued.  Either
way (on queue or drafted) the user is "signed up" to the raid.

Signed-off-by: Douglas Wagner <dwagner@rgare.com>