Bugfix: Fixed issue with all links generating a "new page" when clicked.
Due to code for clicking user links and opening armory in a new window all
links got the same code, meaning every link in the software was opening
a new browser window to load the page.

Also added a title to the Armory Popup of "User@Server" to designate what
is being looked at.

Signed-off-by: Douglas Wagner <dwagner@rgare.com>